+++ src/card/CardAddButton.tsx
@@ -10,7 +10,7 @@ import { SquaresPlusIconOutline } from '@neo4j-ndl/react/icons';
 const NeoAddNewCard = ({ onCreatePressed }) => {
   return (
     <div>
-      <Card style={{ background: '#e0e0e0' }}>
+      <Card className='n-bg-dark-neutral-text-weak'>
         <CardContent style={{ height: '429px' }}>
           <IconButton
             aria-label='add report'