+++ src/report/ReportQueryRunner.ts
@@ -5,6 +5,7 @@ export enum QueryStatus {
   NO_QUERY, // No query specified
   NO_DATA, // No data was returned, therefore we can't draw it.
   NO_DRAWABLE_DATA, // There is data returned, but we can't draw it
+  WAITING, // The report is waiting for custom logic to be executed.
   RUNNING, // The report query is running.
   TIMED_OUT, // Query has reached the time limit.
   COMPLETE, // There is data returned, and we can visualize it all.